Refactor HTMLElement
Also fix a typo in "tabindex" as attribute name.
