+++ lib/jsdom/living/nodes/ElementCSSInlineStyle.webidl
@@ -1,6 +1,6 @@
 [NoInterfaceObject]
 interface ElementCSSInlineStyle {
-  [SameObject, PutForwards=cssText] readonly attribute CSSStyleDeclaration style; // TODO: implement in ElementCSSInlineStyle-impl.js instead of in HTMLElement-impl.js
+  [SameObject, PutForwards=cssText] readonly attribute CSSStyleDeclaration style;
 };
 
 HTMLElement implements ElementCSSInlineStyle;
