+++ lib/jsdom/living/nodes/HTMLElement-impl.js
@@ -2,31 +2,20 @@
 const { mixin } = require("../../utils");
 const ElementImpl = require("./Element-impl").implementation;
 const MouseEvent = require("../generated/MouseEvent");
-const DOMStringMap = require("../generated/DOMStringMap");
+const ElementCSSInlineStyleImpl = require("./ElementCSSInlineStyle-impl").implementation;
 const GlobalEventHandlersImpl = require("./GlobalEventHandlers-impl").implementation;
-const focusing = require("../helpers/focusing.js");
-const conversions = require("webidl-conversions");
+const HTMLAndSVGElementSharedImpl = require("./HTMLAndSVGElementShared-impl").implementation;
 const { isDisabled } = require("../helpers/form-controls");
 
 class HTMLElementImpl extends ElementImpl {
   constructor(args, privateData) {
     super(args, privateData);
+    this._initHTMLAndSVGElement();
+    this._initElementCSSInlineStyle();
     this._initGlobalEvents();
-
-    this._tabIndex = 0;
 
     this._settingCssText = false;
     this._clickInProgress = false;
-
-    this._style = new this._core.CSSStyleDeclaration(newCssText => {
-      if (!this._settingCssText) {
-        this._settingCssText = true;
-        this.setAttribute("style", newCssText);
-        this._settingCssText = false;
-      }
-    });
-
-    this._dataset = DOMStringMap.createImpl([], { element: this });
   }
 
   // Add default event behavior (click link to navigate, click button to submit
@@ -49,32 +38,6 @@ class HTMLElementImpl extends ElementImpl {
     }
 
     return outcome;
-  }
-
-  focus() {
-    if (!focusing.isFocusableAreaElement(this)) {
-      return;
-    }
-
-    const previous = this._ownerDocument._lastFocusedElement;
-
-    focusing.fireFocusEventWithTargetAdjustment("blur", previous, this);
-    this._ownerDocument._lastFocusedElement = this;
-    focusing.fireFocusEventWithTargetAdjustment("focus", this, previous);
-
-    if (this._ownerDocument._defaultView._frameElement) {
-      this._ownerDocument._defaultView._frameElement.focus();
-    }
-  }
-
-  blur() {
-    if (this._ownerDocument._lastFocusedElement !== this || !focusing.isFocusableAreaElement(this)) {
-      return;
-    }
-
-    focusing.fireFocusEventWithTargetAdjustment("blur", this, this._ownerDocument);
-    this._ownerDocument._lastFocusedElement = null;
-    focusing.fireFocusEventWithTargetAdjustment("focus", this._ownerDocument, this);
   }
 
   click() {
@@ -120,13 +83,6 @@ class HTMLElementImpl extends ElementImpl {
     this.setAttribute("dir", value);
   }
 
-  get style() {
-    return this._style;
-  }
-  set style(value) {
-    this._style.cssText = value;
-  }
-
   _attrModified(name, value, oldValue) {
     if (name === "style" && value !== oldValue && !this._settingCssText) {
       this._settingCssText = true;
@@ -137,22 +93,6 @@ class HTMLElementImpl extends ElementImpl {
     }
 
     super._attrModified.apply(this, arguments);
-  }
-
-  get dataset() {
-    return this._dataset;
-  }
-
-  // TODO this should be [Reflect]able if we added default value support to webidl2js's [Reflect]
-  get tabIndex() {
-    if (!this.hasAttribute("tabindex")) {
-      return focusing.isFocusableAreaElement(this) ? 0 : -1;
-    }
-    return conversions.long(this.getAttribute("tabindex"));
-  }
-
-  set tabIndex(value) {
-    this.setAttribute("tabIndex", String(value));
   }
 
   get offsetParent() {
@@ -185,7 +125,9 @@ function callEventBehaviorHook(event, name, targetOverride) {
   }
 }
 
+mixin(HTMLElementImpl.prototype, ElementCSSInlineStyleImpl.prototype);
 mixin(HTMLElementImpl.prototype, GlobalEventHandlersImpl.prototype);
+mixin(HTMLElementImpl.prototype, HTMLAndSVGElementSharedImpl.prototype);
 
 module.exports = {
   implementation: HTMLElementImpl
