+++ lib/jsdom/living/nodes/HTMLElement.webidl
@@ -4,14 +4,16 @@ interface HTMLElement : Element {
   [Reflect] attribute DOMString lang;
 //  attribute boolean translate;
   attribute DOMString dir;
-  [SameObject] readonly attribute DOMStringMap dataset;
+// Implemented in HTMLAndSVGElementShared;
+//  [SameObject] readonly attribute DOMStringMap dataset;
 
   // user interaction
   [Reflect] attribute boolean hidden;
   void click();
-  attribute long tabIndex;
-  void focus();
-  void blur();
+// These three are included in HTMLAndSVGElementShared.
+//  attribute long tabIndex;
+//  void focus();
+//  void blur();
   [Reflect] attribute DOMString accessKey;
 //  readonly attribute DOMString accessKeyLabel;
 //  attribute boolean draggable;
@@ -22,6 +24,7 @@ interface HTMLElement : Element {
 };
 HTMLElement implements GlobalEventHandlers;
 HTMLElement implements ElementContentEditable;
+HTMLElement implements HTMLAndSVGElementShared;
 
 // https://drafts.csswg.org/cssom-view/#extensions-to-the-htmlelement-interface
 partial interface HTMLElement {
